server/requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
ecdsa==0.18.0
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
install==1.3.5
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
